libapp: add initialize() method for App::Object
Due to the unfortunate implementation of shared_from_this(), objects
cannot call the function during construction. This means that a second
pass where this function can be called is needed, as otherwise objects
will not be properly constructed. Before now, the attempted solution
was to lazily initialize things if possible, but this did not always
work.