kernel: refactor the sending of packets to allow for a send queue
Now the kernel will reserve its own "header" around any sent packet,
which includes space for a linked list of other packets. This will
allow the sending of packets to be queued by the kernel or a network
interface. This structure is the same one used for the recieve packet
queue, which means that data sent through the loopback interface can
be put into the queue with 0 additional overhead.