kernel: improve the packet layering mechanism
Rather than allocating enough contiguous space to hold the finished
packet up front, the networking code works with a flexible packet
structure, which is basically an array of buffers with a specific
type identifier.

The TCP code now puts its TCP header in the third header entry of
this packet structure, and does not worry at all about the ip or
link layer headers. This change should make it a lot easier to use
ipv6, or to use some form of packet encapsulation such that TCP is
being sent of something more than just ipv4.

This patch simply implements the mechanism, so there is still work
to be done to make things more functional.