libdl: implement dladdr()
This function is useful for generating debug backtraces when an
application crashes, and for doing general introspection. Currently,
the value of the dli_fname field is incorrect, since it should point
to the full path of the shared library instead of the name of the file.
This should be fixed when the dynamic loader supports libraries in
paths other that /usr/lib.