kernel: add init function mechanism
This patch begins to reduce the number of explicit init_* calls in
the kernel (which fill up almost all of the kernel's entry code).
The mechanism is exactly like the .init_array ELF section, but there
is more than one list, so that the order can be more precisely
controlled. The next patch should actually use this mechanism in
the proper places, and potentially add more levels so that adding
new code isn't excessively difficult.