kernel: run most of the init code in process context
This patch moves the kernel idle tasks to PID 0, and reserves PID 1
for the "proper" init task. This is now the task which exec's the
userspace /bin/start program.

By making the init process run in an actual scheduling context,
initialization code will be able to do things like wait for IRQs or
sleep without any difficulties.

The only issue right now is that drivers would like to use timer
features during device initialization, but the presence of timers
is detected at the same time as other devices. This likely can be
fixed with a two pass approach, and would most likely prove to be
quite a good idea.