kernel: add /proc/blockdev file which lists all block devices
This providese useful information, and the list of block devices
will be used by the kernel (eventually) to determine which one to
mount as root (according to the kernel command line) during the
init process. This is necessary to ensure that the kernel can boot
properly on partitioned devices, especially when other OS's are
installed.