kernel: move signal dispatch out of the scheduler
Instead of the scheduler unblocking a blocking task with pending
signals, the act of queuing a signal now unblocks the task.

In order for this to work, tasks must check if they have pending
signals before blocking (if the wait is interruptible), since the
scheduler isn't there to subsequently unblock this now blocking
task.

Currently, this mechanism should be completely broken on SMP, where
there a myriad of issues with the current implementation. Having
__kernel_yield() return directly through the task state's rax is a
problem since a task can set its state to WAITING before it is no
longer running, causing any writes to its task state to be discarded.
Additionally, concurrent wake ups must be protected against using a
spin lock. And signal dispatch was never made SMP safe, and so also
needs a spin lock to protect against concurrent dispatch.

Additionally, this patch should also break SIGCONT functionality, as
it will never be dispatched to begin with. This is fine because POSIX
says SIGCONT's functionality should occur when the signal is generated,
not when it is handled.