libipc: add macro generator for IPC messages
Messages are defined like so:
IPC_MESSAGES(Name,
    (M1,
        (int, a),
        (int, b)
    ),
    (M2,
        (long, x)
    )
)

which creates a enum class Name::Type and POD structs
M1 { int a; int b; } M2 { long x; }, as well as
serialization/deserialiation methods for sending
such structures over sockets.

To do so, the C preprocessor is tricked into recusively
expanding a macro to accomplish argument iteration.
Arguments are grouped in parenthesis in order for this
iteration to work at the right level.

This is certainly a terrible idea, but at the same time
is quite satisfying (much like LIIM::Variant).

Credit to this post for recursive macro expansion:
https://github.com/pfultz2/Cloak/wiki/C-Preprocessor-tricks,-tips,-and-idioms