toolchain: fix misc errors
This additionally requires building libc.so before libgcc, as for
whatever reason libgcc_s.so.1 needs to link against libc. This is
fairly annoying, but hopefully can be fixed by implementing
dl_iterate_phdr, as that will allow for a better exception handling
mechanism which does not require libgcc registration calls.