edit: fix crash when pressing PAGEUP
The move_cursor_up() function mistakenly thought that m_row_offset
has to be at least greater than 0. This is not true, because the
cursor does not necessarily have to be currently in view. If the
cursor is hidden above the current view, this field will be negative.

Thus the cursor cannot be moved up when cursor_row() + m_row_offset == 0,
not when both these fields are equal to 0.