kernel: refactor mount points to be simpler and more correct
This patch changes parent inodes from having a separate list of mount
objects to having a single mount object per child inode. The helper
function maybe_cross_mount_point() converts this "hidden" inode to
the root inode of the file system mounted on it.

Doing so makes code no longer have to consider mount points separately,
and instead things just work out.

In addition, fs_do_mount() is now much simpler, since path lookup
can be done with a regular iname() call.

For now, the initrd which is mounted as root earlier in the boot
procecss is removed when a hard disk is found to boot from. At this
point, a new root is instantiated. Then, the kernel mounts the initrd
a second time.

Because the umount() operation is not implemented, the removal of the
first initrd file system leaks every single inode in the initrd (~50).
This can be fixed by simply implementing umount() or, since the initrd
is read-only, returning the same super block each time the initrd is
mounted (this is already what devfs does).