kernel: fix the local APIC timer for real
The reason the timer wasn't firing was that it was incorrectly
set to one-shot mode instead of periodic mode due to a typo.

Fixing this issue caused the APIC timer to fire ridiculously often,
though. This was due to both the improper calibration and improper
use of this calibration.

Luckily, the fix was the same for both issues, since the problem was
with the use of the time_divide function. All callers were using
this function to divide the wrong way. The correct division is to
divide the base frequency of the timer by the frequency corresponding
to a passed in interval. Doing this gives the correct number of ticks
to use.

Before, the code divided the interval by the base frequency. This
was completely wrong.