kernel: do a better job handling mount points in the VFS
This patch adds a new iname flag: INAME_DONT_FOLLOW_TRAILING_MOUNT_POINT.
This causes the iname() to return the "shadowed" inode instead of the
mounted root in cases where a mount point is directly looked up. This
is specifically useful for the mount() function, which cannot allow
multiple mount points to stack.

Additionally, fs_default_dir_read() now returns root inodes in stead
of the "shadowed" inodes when reading a directory. This improves the
correctness of the "ls" program.