kernel: properly free all inodes when unmounting tmpfs
For now, this works by storing all inodes in a linked list
on the super block. This is somewhat of a bottleneck for inode
creation, but at the same time, some serialization is already needed
to allocate an inode identifier. As such, this linked list is
probably acceptable long term.