kernel: begin refactoring ext2fs to use a proper block iterator
Current code reads from inodes in an extremely ad-hoc fashion, which
meant that triply indirect blocks were completely unsupported, and
any form of indirection for directory entries was ignored.

This patch introduces a ext2_block_iterator structure which handles
all forms of indirection in a correct and reuse-able fashion.

For now, this only works for fully iterating an inode, and doesn't
support allocating new blocks for writing or starting the iteration
not from the beginning. Once these work, this mechanism can be used
in all locations that read inodes.