kernel: convert ext2_truncate to use ext2 block iterator
At long last, the ext2fs driver actually works with files of any
size.

Unfortunately, there are still some bugs with driver. The most
obvious is that there are plenty of race conditions when
allocating/deallocating blocks.

In addition, symlink files need special treament because of their
inline storage, which doesn't always happen in the code.

But for the most part the file system driver now is a lot more
stable.