kernel: fix ext2fs block and inode bitmap functions
The code mistakenly thought that the raw data it was writing was of
type uint64_t* instead of void*, so it incorrectly adjusted the
pointer offset when it was uneeded. This patch means that the
ext2fs driver now does not ruin the whole file system whenever
it update these bitmaps.