edit: move most edit classes into the libedit library
The core edit functionality (the Document and Command objects), can
now be used by anything in the system.

Due to poor git management, this commit also adds some features to
libedit. Namely, the display is allowed to have a different width
per line, and the SingleLine mode is expanded to allow for multiline
editing while still being submittable and not backed by any real file.