libtinput: begin working on a generic REPL library
This commit introduces a simple abstract Repl class, which is meant
to be subclassed in applications like the shell.

In addition, libedit is used in combination with a custom ReplPanel.
As of right now, features like history and tab completion are absent,
but it is possible to have syntax highlighting and proper multiline
editing in addition to simple prompts.

This will hopefully drive features in the edit program, like
autocompletion and word wrapping, which then can be easily implemented
in the Repl.