libtinput: handle terminal escape sequences in prompts adequately
Knowing the number of characters that will be printed by a terminal
is almost impossible to do accurately. For now, a naive approach
is used.

Most likely the correct solution would involve querying the terminal
to get the current cursor positions, which would allow the terminal
to be the one doing the measuring.