libedit: add initial support for tab autocompletion
This patch adds an editor mode which makes all tab presses
try to get autocomplete suggestions, and use them if only a single
one is suggested.

It is up to the Panel to render the suggestions in some way if they
are multiple that the user must choose from. Additionally, the default
implementation never returns any suggestions (but this should eventually
be integrated by the document type mechanism).