libtinput: add auto complete support to the Repl
The actual Repl subclass is the one which comes up with the suggestions,
and then the ReplPanel informs the document about them.

The ReplPanel automatically converts multiple suggestions into a
single one if they all share some common prefix. This is standard
behavior for the shell, but may need to be changed in other use cases.