libtinput: add simple History implementation
The History class is a simple wrapper around a Vector<String>, and
is used to query history information.

The ReplPanel will cycle the history when up/down is pressed only on
the first/last line of the document. This way multi-line editing is
still supported.

To cycle the history, the ReplPanel replaces its current Document with
a new one created from the history item string. These documents are
cached once used in a member variable.