libsh: fix ShLexer's handling of io redirects
The lexer previously failed to work on input like:

echo "Hello, World!" >nice=file

This is because nice=file while lexed as an ASSIGNMENT_WORD and
not a regular WORD. This commit fixes this issue, and moves the
contents of a here document into its own unique token. In addition,
the is_io_redirect() function is fixed to actually consider all the
possible io redirect tokens. All these fixes make the lexer much more
accurate at determining whether or not a given WORD can be an
ASSIGNMENT_WORD.