libtinput: read/write history to a file
In addition, History::add() does not actually add the history text
if the command is empty, starts with a space, or is the same as the
previous history text.

Instead of converting any multiline input into a single line, this
implementation serializes multiline history entries by prepending
a space in front of each subsequent line. This is totally fine
because any text that starts with a space simply won't be added.

This menas that long multiline input strings will be remembered
that way if they are scrolled to.