libtinput: handle mouse events in the shell if they are sent
Note that although its very cool to navigate the shell input with
a mouse, doing so prevents the user from copy/pasting command output
in the terminal. Since I'm not sure what to do about this, the mouse
is still set to disabled even though ReplPanel can handle mouse events.