kernel: fix bug when delivering terminating signal
Ironically, this bug was introduced by commit
97b53c33e8e6a0f9cceecec0c23c026c4e6ecb44, which claims to have
fixed a bug when delivering a terminating signal. It instead causes
the kernel to crash anytime such a signal is delivered, by triggering
a double free. Namely, the task was exited and freed twice.

This patch fixes this by using the correct function
task_set_state_to_exiting() which sets the tasks state in the correct
fashion, and includes sanity checks around exiting a task twice.