libc: make the libc strcmp() function not broken
This was the same bug found in strncmp() a week ago. Luckily this
bug didn't lead to buffer overrun, since the NULL terminator
is the smallest possible character value. As such, the loop body
would immediately return.

The code is actually only incorrect if s2 had newline in it, which
fortunately is very rare. Now this is fixed.