libliim: refactor Generator to allow co_yielding other Generators
This mechanism allows Generator functions to be composable and call
other Generator functions. Unfortunately, this implementation does
not work if the depth of recursive co_yields is greater than 2.

The main problem is that I do not know how to get the return object
a promise already created. As such, it makes writing the Generator
code extremely difficult, in addition to the complexity already
introduced by coroutines. This will hopefully be improved later.