libext: begin writing a class for deflate Compression
This uses the new Generator<> mechanism to allow the compression
to be streamed and fully resumable. Once the compressor runs out of
data, if its not the last block, it will co_yield, which will then
allow the caller to supply more data. The same mechanism will also
let the caller expand/drain the output buffer if needed.

This is a proof-of-concept for the Generator<> class, which can
hopefully also be used in the Decompressor as well, which would
get rid of the ugly goto/state machine it currently uses to be
resumable.