libliim: add Byte{Reader,Writer} helper classes
These classes provide common implementations for functions which need
to read/write to a sequence of bytes.

The ByteWriter currently owns a ByteBuffer, which it then uses to
insert its data. I would prefer that it took a Span like ByteReader,
and wrote its data there, but there are some problems. Mainly, there
is no way for the ByteWriter class to easily signal how many bytes it
has written, and it has no knowledge of its underlying ByteBuffer, which
makes things less safe.