kernel: fix bug when cloning page tables for fork()
Disabling pre-emption was not enough because the current cr3 is
changed from the value stored in the process's arch_process struct.
As such, any interrupted process would lose the new cr3 value, and
then things would break.

In addition, since this operation modifies the kernel address space,
it is a bad idea to let irqs run in that context, even if this issue
would be avoidable. As such, its best to just hard disable irqs during
the page table manipulations.