kernel: fix horribly incorrect code in the user mutex implementation
The code was lazily refactored when wait_queue_entry was first
introduced, and so the function to enqueue a task onto a user
mutex took a task*. Then it constructed a wait_queue_entry on the
stack, and enqueued it, and then returned. So now the wait_queue_entry*
is no longer valid.

In addition, the code never removed the wait_queue_entry from the
wait_queue upon being woken up, which meant that once the system
call returned, even a properly stack allocated wait_queue_entry would
go out of scope. The wait_queue_entry does not automatically remove
itself because the task usually also needs to evaluate a condition
to know if it should really wake up. If it shouldn't wake up, it
never even removes itself from the queue, so it can go right back
to sleep. I believe this also makes implementing the select and poll
system calls easier, since they cause a task to enqueue itself on
multiple queues, and likely only 1 will actually cause it to wake
up. In this case, the task has know way of knowing which queues it
actually needs to remove itself from.

The fix is to simply change add_to_user_mutex_queue to take a
wait_queue_entry* and enqueue that. In addition, the added
wait_queue_entry is not properly removed. Note that this code need
not check for any conditions because it is userspace who is the one
who actually writes to the mutex, and so it is userspace you tests
that the lock is actually unlocked.