kernel: don't clobber a task's sched_state in the page fault handler
Because inode's are currently synchronized using mutexes, the page
fault handler needs to be able to wait on said mutexes when handling
certain types of page faults. To do so, the task must be running in
kernel mode.

Previously, the task may or may not have been in kernel
mode, and waiting on a mutex sets the task->sched_state but not the
task->in_kernel field. Then, the task->sched_state was not properly
restored.

This meant that the sched_stress_test program, which simply
loops forever, would have its sched_state clobbered from INTERRUPTIBLE
to UNINTERRUPTIBLE whenever it fauled. And since it never made any
system calls, the kernel could never "fix" this. As such, the scheduler
refused to deliver any signals to the task, allowing to run forever
and be unkillable.

This is fixed by making sure to properly handle the task struct
fields on page fault entry/exit. Note that any other handlers
which do things other than crash the process or deliver a signal
will also need to perform these actions.