libext+libliim: make ByteWriter operate on a fixed size Span<uint8_t>
Now that the caller ultimately owns the output store, the
encoder/decoder classes don't have to worry about it. Additionally,
the ByteBuffer will never be resized.

Another nice thing is that there is a lot of duplication in each
of the Encoder/Decoder classes. This will soon be refactored into a
common base class, which will clean things up even more.