libext: refactor DeflateDecoder to use the new StreamDecoder mechanism
Note that the large FixedArray member variables of DeflateDecoder are
not moved into the decode() coroutine, as saving these values would
be excessively expensive. But all the local variables like m_i can now
be put into the coroutine itself.

Because of Generator<>'s limitation of only one recursive call to a
Generator<> function, everything needed to be shoved into the decode()
function. Things would be a lot better if this functino could be broken
up, but this isn't possible yet.