kernel: fix bug in __time_wakeup_after and time_wakeup_after
If the user specified timeout for pselect(2) or ppoll(2) is 0, then
the kernel should not block at all. Previously, the kernel would setup
a timer to fire immediately, which would then wakeup the blocking
process. However, because the timer's value was 0, it never fired,
even though it was added to the clock's list of pending timers.

Because of this, the stack allocated timer in the wakeup_after functions
was never removed from the list, which then became corrupted. This
is easily fixed by simply checking for a 0 timeout.