libedit: map ALT+d to delete word (CONTROL+DELETE)
This is necessary solely because the vscode integrated terminal
decided to send ALT+d (ESC d) instead of CONTROL+DELETE (CSI 3;5~),
when CONTROL+DELETE is actually pressed. This is almost certainly
because of bash's terrible default key bindings, since ALT+d is
mapped to the function CONTROL+DELETE should be mapped to by
default.

To support this work around, the text editor must make ALT+d
delete a word.