window_server: add button which spawns a new terminal
Now this system is not completely useless once the terminal is closed.
Unfortunately, the terminal is running under the ws user, so this is
horribly broken.

The way to fix this is to move the taskbar outside of the window server
and into a separate process. This is nice it will allow the taskbar
to use the full range of libapp widgets, and will hopefully simplify
the window server's logic a bit.