kernel: "fix-up" the ptmx code some more
Although ugly, the code now properly uses the locks it should when
updating the various file state objects.

In addition, the VEOF character is not inserted directly into the
input buffer, which the kernel will then turn into an EOF once
the stream reaches that character. Before, the VEOF character was
essentially ignored, which made cat hang once ^D was pressed.