libc: add backtrace() and dump_backtrace() functions
These are in the execinfo.h header, which mirrors the fact that the
backtrace function is the same as glibc's. The dump_backtrace()
function is specific to this libc.

Note that dump_backtrace() gets its symbol information from the
dynamic loader, so no symbols will be available for statically
linked executables. In addition, applications must be compiled/linked
with -rdynamic in order for non-library symbols to be visible. Even
with this option, static functions will not be visible.

The backtrace() function requires -fno-omit-frame-pointer to work
properly, since it uses the frame pointers to do the backtrace. It
at least ensures its pointers are valid, so its always safe to call.

It would be better to read the DWARF tables instead, but doing so
is obviously more complicated. In addition, its possible that some
of the exeception unwinding code in libgcc could be used, but this
would require more research into its API.