kernel+libc: add the UMessge Watch protocol
This mechanism allows a process to monitor the state of multiple
paths, exposed through a socket connection to the kernel. This makes
things like automatically updating a file manager or hot-reloading
configuration files possible.

For userspace, the socket can created with the UMESSAGE_WATCH protocol,
and then paths can be watched or unwatched through umessage requests.
Changes are then reported through readable umessage events on the created
socket.

Note that each watcher is assigned a unique identifier in userspace,
which userspace can then use to refer to it later. This is like the
watch descriptor concept from linux's inotify, but the identifiers
are allocated by userspace and not the kernel.

This is mostly done because its not easy to pass a watch descriptor
back to userspace without adding another umessage event, which would
then make things asynchronous. Work arounds include overloading the
return type from the write function, some sort of extra data from
sendmsg, or passing a pointer in the event that would be filled by
the kernel. The last option is the most reasonable, but allowing userspace
to allocate its own descriptors also seems fine.

The drawbacks are that queued_umessage's become unique for every watcher,
since each watcher has an arbitrary identifier. If the kernel assigned
them, it could use something like the inode's index and inode's device
to create a unique identifier, which would allow each umessage to be
shared.

In addition, userspace can assign duplicate identifiers if it wants to,
and the kernel will just let userspace missuse the API. This is unfortunate
but doesn't seem like a security vunerability, so it should be fine.

For now, the only supported event is that the inode's content changed,
and it is only reported by the ext2fs. Events for directories will
eventually be added, as will event generation for other file systems.