libeventloop: add FileWatcher class
This is C++ wrapper around the newly created UMessage Watch API.
The event reporting mechanism is very simple, but will be expanded
when the kernel report more events.

Note that this class is currently exclusive for the code running on
the OS. If a portable application like the shell or the text editor
want to use this, support for linux's inotify will probably be added.