libeventloop: allow enqueuing events from different threads
This makes the event queue thread safe by protecting it with a mutex.
In addition, the event loop will signal itself whenever events are
queued from a different thread. This ensures the main thread will
wake up and process the events.

This makes it so that event loop programs can spawn other threads and
recieve data from them.