libapp: support creating basic windows using SDL
This patch moves some of the old App::Window functionality into a
App::PlatformWindow class, which is an abstract class with 2 different
implementations. In addition, the global HashMap of window ids to windows
is now protected with a mutex.

Note that only a few SDL events are currently handled, so things like
focus, key events, and mouse events are ignored for now.