libedit: compute the rendered line contents before drawing
For now, the rendered line is computed directly every time a line
is rendered, but this could easily be cached. Doing so makes the
code which actually draws to a panel much simpler, which is important
if the panel were to be doing some complicated drawing.

In addition, adding some more information to the rendered line will
allow lookups from visible panel positions to indexes into a line
to be performed quickly and accurately.

This change reveals that the way in which CharacterMetadata is stored
is quite silly. This field should clearly be some sort of range, and
is does not need to be stored on a per-character basis. This requires
the rendered line to also maintain a CharacterMetadata per rendered
character, which further wastes memory.