libedit: add a Cursor class
This is in preparation for making the Document's cursor relative
to line indices instead of its display position. Doing so ensures
the cursor's position is always valid, and allows for the Command
classes to work on a line directly. This should lead to a substantial
simplification of the code, and prepare for the possibility of having
multiple cursors.