libedit: fix auto complete preview
To correctly locate the cursor, the Line class must know whether the
rendered text corresponds to a real line index of not. This is done
by tracking the rendered span for each character in the line. Since
the auto complete preview will have no span, the Line will ignore it
when iterating metadata. The position converter functions consider the
preview to be part of the previous span, which ensures the cursor is
drawn in the correct position.