libedit: implement preliminary word wrapping
This patch implements word wrapping in the Line class. This involves
lots of complicated calculations when converting positions into indices,
due to the code lacking a proper data structure for rendered lines. This
is clearly demonstrated by the fact that 5/6 of the Line classes member
variables are used for storing rendered information, and all of them
must be cached. Without this basic caching strategy, the text editor
is completely unusable (with word wrapping on), so these optimizations
had to be included in this commit.

There are still lingering issues with this implementation, concerning
especially the panel rendered line counts, as well as various issues
with integrating with ReplPanel.