terminal: send ASCII 8 instead of ASCII 127 for KEY_BACKSPACE
This patch additionally ignores any ASCII 127 characters it recieves.
libtinput expects ASCII 8 for this functionality, so its better for
the terminal to just send to correct value. Previously, whenever
backspace was pressed, ASCII 127 was echoed back to the terminal,
which then went into an infinite loop trying to display it, due to
the fact that or'ing it with 0x40 didn't stop it from being a control
character.