libedit: convert Selection to use TextIndex
Because Selection was integral to most Commands, this patch also
converts the Command class'es to work with TextIndex as well. The
Document::move_cursor_to() now takes a text index to make things
easier.

This greatly simplifies the Selection class, since normalizing it
is now trival (thanks to TextIndex's comparison functions).