libedit: add Panel::inject_inline_text_for_line_index()
This allows the ReplPanel to inject its prompts directly into the
Line's rendering algorithm. However, this menas that TTY escape
sequences do not work, as their length is not correctly measured.
In addition, inline text handling is implemented somewhat hackishly,
because of how where the cursor should be drawn. For a prompt text,
the cursor should go afterwards, but for an auto complete preview,
the cursor goes before the inline text. This logic is hardcoded
into line by special casing index 0, but this should be made more
general.