terminal: add support for Origin Mode
This mode offsets the row numbers by the start of the scrollable
region (as set with CSI Pt ; Pb r), and additionally forces the
cursor to be within this region.

This patch also resets the cursor and margins when changing between
80 and 132 column mode, and makes tab characters move the cursor to
the end of the current line.

Doing so makes the terminal "pass" vttest section 1 (cursor movement).