terminal: support truecolor
This patch implements:
  CSI 38 ; 2 ; RED ; GREEN ; BLUE to set the foreground, and
  CSI 48 ; 2 ; RED ; GREEN ; BLUE to set the background.

Some terminals also support a ':' as a delimiter, but since this
would require changing the TTY escape sequence parser and is generally
less widely supported, it does not need to be implemented now.

In addition, the terminal now sets the TERM enviornment variable
to 'xterm-256color', which advertises support for this feature.