libedit: improve handling of multiple cursors
This patch makes MultiCursor store all of its Cursor objects in
consecutive order, and disallows any selection's to overlap. This
has numerous advantages.

  - Firstly, most editing functions will now work properly by simply
    iterating the Cursor's in reverse order. Previously, inserting or
    deleting characters had random result based on the current ordering
    of the MultiCursor.
  - Secondly, this makes checking for duplicate cursors O(n) instead of
    O(n^2), as well as simpler to implement. However, because the
    MultiCursor uses a Vector<> internally, removing cursors is still
    somewhat expensive.

Note that MultiCursor never actually checks that this order is enforced
after the insertion of a Cursor. As all cursor movement functions are
applied to all cursors, never only one, there should be no way that
this property could be invalidated. If there were, MultiCursor would
need a way to re-sort() the cursors.