libedit+edit: make cursor rendering part of CharacterMetadata
The CharacterMetadata class now indicates if the current position
should be rendered as the cursor or not. This makes multicursor
rendering much more straight forward for all classes.

In order to support this, the TextRangeCollection now allows indexes
equal to the line's length, and LineRenderer generates a trailing space
character for the cursor to be rendered at (if it is at the end of the
line). This notably causes line wrapping to occur for lines which
previously fit exactly on the panel. This behavior is somewhat strange,
but so was the alternative, which misplaced the cursor.