sh: add the getopts builtin
This implementation uses the getopt() library function, but it most
likely shouldn't. This is because getopt() will maintain global state
which requires the argc/argv passed to it to be stable. Espacially if
the arguments are passed directly into getopts, this will not be case,
and so the getopts command will completely break.