libliim: refactor StringView to have a much less insane implementation
Previously, StringView stored a start and end pointer, and the end
pointer was inclusive instead of exclusive. Now, StringView stores
data and size members, end() points to the exclusive end (so now
is a proper end iterator), and the constructor expects the any
start/end pointer pair it gets have an exclusive end.