libtest: add method for spawning a subprocess
This spawns an arbitrary subprocess and waits for its exit code. It
allows for a function to run just before execve() is called, which
allows programs to perform certain setup actions.

This is mostly useful when testing certain behaviors of the kernel,
in particular, process attributes which are preserved/cleared across
an execve() system call.