libtui: begin work on a tui framework
This works similary to libapp, but is designed for terminal based
UIs instead of GUIs. This will be used mainly be the edit program
to faciliate displaying multiple documents simultaneously, as well
as making complicated UI components like popups easier to implement.
In addition, this will be used by the Repl code, which will hopefully
allow some sophistication there as well.

This initial commit implements a basic layout system, but neglects to
properly dispatch key/mouse events and maintain a notition of a focused
panel.

With this patch, it is clear that the Repl code should be moved from
libtinput into its own separate librepl library.