edit: only clear areas of the panel where no text will be drawn
Previously, the entire area was cleared and then painted over. On
most terminals, this was not an issue, since they wait to repaint
the screen until after they consume everything in the input buffer.
However, the vscode terminal would actually render the intermediate
state, which caused the screen to flicker. This patch fixes that issue
by only clearing areas where no text is displayed.

This could also be solved by buffering terminal updates in libtinput,
but this is a lot more work to implement.