librepl+libedit: add mechanism to choose between multiple suggestions
This shows the suggestions below the current cursor, and one can be
chosen using either up/down or tab/shift+tab, followed by enter.

For now this consumes all key events, but it would be better to
forward events to the ReplDisplay but keep this widget open and
responsive to new key presses.