libtinput: add API for splitting text into glyphs
This API is extremely primitive, but should be suitable for a terminal
based renderer. Currently, the need for glyphs comes from librepl's
ability to specifiy terminal escape sequences in its prompts text,
which must be considered to have a width of zero. In the future, this
API should take care of a lot of unicode stuff.