libliim: begin implementing a c++ based formatting mechanism
This lays out the basic mechanism, which is based of fmtlib and the
c++20 standard.

The implementation revovles around the FormatArgs,
which are type erased by storing only a const void* and a c style
function pointer. This function pointer is generated by a template
function, which calls Formatter<T>::format(). Since this template
function depends on only a single type, instead of the entire
argument list, the compiler is not forced to generate massive
templated functions. In fmtlib, this type erased paramater is actually
a discriminated union (storing either an integral value or this custom
void* type), but this should be overkill for now.

Instead of using the iterator based approach for the Formatter<T>::format()
function, this implementation will extend the FormatContext to make
the API simpler but less flexible.

This patch implements exactly one formatter, for StringView, and the
entire mechanism does not do anything to actually parse the contents
within {}.

In the future, this will be used to implement some nice logging
facilities as well as nice c++ text output to files.