libc: ignore EOF marker at the beginning of fgetc and fread
Previously, the code optimized for the case where a read had previosly
returned EOF, and short circuited do just return EOF again. However,
this is clearly wrong, particularly in cases where the underlying file
is a terminal. Even if not a terminal, another program can appead to
the file anyway, so an actual read must be performed.

Due to the lack of actual tests for stdio functionality, the short
circuit condition on error remains in place. I expect this is not the
correct behavior, but it is less critical than this EOF condition,
which broke libtinput.