sh: remove explicit management of termios
Since librepl manages the termios itself, there is no need for
the shell to do anything. In particular, the state is restored when
the user presses enter, and set to raw mode before prompting the user
for more input.

This patch makes it so the shell properly handles buffered input.
Any input typed while another command is running will now be processed
by the shell (instead of being discarded), assuming the currently running
application never bothers to read it.