kernel: add a removed event to UMessage Watch
This in it of itself would not be complicated, but since removed inodes
can never generate events again, they also need to somehow remove
themselves. This patch implements the removed event, but doesn't handle
this case exceptionally well.

When an inode is removed, the watch object is removed from the inode's
list of watchers, but it remains in umessage_queue's list of watchers.
It is up to userspace to call remove watch on the inode, which then drops
its inode reference, which then finally frees the inode. Note that holding
an inode reference only stops freeing the object, not removing it from disk
(which has already occured by this point). It would be better if this
removal was automatic, but doing so proved to be quite difficult.

This is because both the umessage_queue and the inode have their own
lists, each of which are protected by different locks. If one side wants
to remove it from the other, it must somehow aquire the other's lock.
Since each side aquires the locks in different orders, deadlock will
occur. Any sort of solution appears to require lots of complex synchronization.

This patch also implements test cases for this functionality. Rather than test
the kernel implementation directly, libeventloop's FileWatcher class is tested
instead, which uses linux's inotify in native builds. Since both implementations
should have tests, it feels better to test both at once, even though testing the
umessage part directly could be useful.